Add files via upload
Movie Scheduling in a Multiplex
This project aims to schedule the given movies to maximise the profit of the multiplex mall. The priority of a movie is calculated and used to assign the screens to the  movie to generate a schedule for one week. This will be done by analysing the number of tickets unbooked and cast weightage, demand score, respectively.